+++ packages/css/src/index.js
@@ -31,7 +31,7 @@ const aliases = {
   py: 'paddingY',
 }
 
-const multiples = {
+export const multiples = {
   marginX: ['marginLeft', 'marginRight'],
   marginY: ['marginTop', 'marginBottom'],
   paddingX: ['paddingLeft', 'paddingRight'],
@@ -39,7 +39,7 @@ const multiples = {
   size: ['width', 'height'],
 }
 
-const scales = {
+export const scales = {
   color: 'colors',
   backgroundColor: 'colors',
   borderColor: 'colors',
@@ -162,15 +162,12 @@ const responsive = styles => theme => {
     }
     for (let i = 0; i < value.slice(0, mediaQueries.length).length; i++) {
       const media = mediaQueries[i]
-      if (value[i] == null) {
-        next[media] = {}
-        continue
-      }
       if (!media) {
         next[key] = value[i]
         continue
       }
       next[media] = next[media] || {}
+      if (value[i] == null) continue
       next[media][key] = value[i]
     }
   }